requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
lightgbm>=4.3
shap>=0.45
ta>=0.11
joblib>=1.4
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
requests>=2.32
beautifulsoup4>=4.12
feedparser>=6.0
vaderSentiment>=3.3
PyYAML>=6.0
APScheduler>=3.10
pyarrow>=16.1

# MongoDB dependencies
pymongo>=4.6
motor>=3.3
